pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kenyan-doc-processor"
version = "2.0.0"
description = "AI-powered document validation and data extraction for Kenyan official documents"
authors = [
    {name = "Gl4d3", email = "gl4d3@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
    "aiohttp>=3.9.1",
    "pydantic-settings>=2.1.0",
    "opencv-python-headless>=4.8.1.78",
    "pytesseract>=0.3.10",
    "Pillow>=10.1.0",
    "numpy>=1.24.3",
    "scikit-image>=0.22.0",
    "PyMuPDF>=1.23.8",
    "easyocr>=1.7.0",
    "pydantic>=2.5.0",
    "python-dateutil>=2.8.2",
    "PyYAML>=6.0.1",
    "redis>=5.0.1",
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "pytest-cov>=4.1.0",
]
dev = [
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

[project.urls]
Homepage = "https://github.com/gl4d3/kenyan-doc-processor"
Repository = "https://github.com/gl4d3/kenyan-doc-processor"
Documentation = "https://github.com/gl4d3/kenyan-doc-processor#readme"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true